Extract dotnet framework locator methods
